tests/krate/publish: Move more tests
