+++ src/tests/krate/publish/readme.rs
@@ -1,4 +1,4 @@
-use crate::builders::PublishBuilder;
+use crate::builders::{CrateBuilder, PublishBuilder};
 use crate::util::{RequestHelper, TestApp};
 
 #[test]
@@ -52,4 +52,40 @@ fn new_krate_with_readme_and_plus_version() {
         "readmes/foo_readme/foo_readme-1.0.0+foo.html",
     ];
     assert_eq!(app.stored_files(), expected_files);
+}
+
+#[test]
+fn publish_after_removing_documentation() {
+    let (app, anon, user, token) = TestApp::full().with_token();
+    let user = user.as_model();
+
+    // 1. Start with a crate with no documentation
+    app.db(|conn| {
+        CrateBuilder::new("docscrate", user.id)
+            .version("0.2.0")
+            .expect_build(conn);
+    });
+
+    // Verify that crates start without any documentation so the next assertion can *prove*
+    // that it was the one that added the documentation
+    let json = anon.show_crate("docscrate");
+    assert_eq!(json.krate.documentation, None);
+
+    // 2. Add documentation
+    let crate_to_publish = PublishBuilder::new("docscrate", "0.2.1").documentation("http://foo.rs");
+    let json = token.publish_crate(crate_to_publish).good();
+    assert_eq!(json.krate.documentation, Some("http://foo.rs".to_owned()));
+
+    // Ensure latest version also has the same documentation
+    let json = anon.show_crate("docscrate");
+    assert_eq!(json.krate.documentation, Some("http://foo.rs".to_owned()));
+
+    // 3. Remove the documentation
+    let crate_to_publish = PublishBuilder::new("docscrate", "0.2.2");
+    let json = token.publish_crate(crate_to_publish).good();
+    assert_eq!(json.krate.documentation, None);
+
+    // Ensure latest version no longer has documentation
+    let json = anon.show_crate("docscrate");
+    assert_eq!(json.krate.documentation, None);
 }
